created a components directory, created a navigation directory in components, moved navbar and footer to the navigation directory. Adjusted navbar and footer to wrap children in layout.tsx
