+++ src/app/layout.tsx
@@ -2,8 +2,8 @@ import './globals.css'
 import type { Metadata } from 'next'
 import { Inter } from 'next/font/google'
 
-import Navbar from './navbar';
-import Footer from './footer';
+import Navbar from './components/navigation/navbar';
+import Footer from './components/navigation/footer';
 
 const inter = Inter({ subsets: ['latin'] })
 
@@ -19,9 +19,11 @@ export default function RootLayout({
 }) {
   return (
     <html lang="en">
-      <Navbar />
-      <body className={inter.className}>{children}</body>
-      <Footer />
+      <body className={inter.className}>
+        <Navbar />
+        {children}
+        <Footer />
+      </body>
     </html>
   )
 }
