--- src/app/navbar.tsx
@@ -1,9 +1,0 @@
-
-const Navbar = () => {
-
-    return (
-        <h1>Navbar</h1>
-    )
-}
-
-export default Navbar;
